4W07 Repeat can be removed from the core
