+++ lib/shell.js
@@ -6,6 +6,38 @@ Fiber.prototype.macro = function(f) {
     return this;
 };
 
+// Repeat a child fiber. The optional delegate method `repeatShouldEnd`
+// should return true when the repetition should end and is called before
+// every iteration with the current iteration count (starting at 0 before
+// the first iteration begins).
+Fiber.prototype.repeat = function(f, delegate) {
+    const body = new Fiber();
+    this.
+        sync((fiber, scheduler) => {
+            if (delegate?.repeatShouldEnd?.call(delegate, 0, fiber, scheduler)) {
+                return;
+            }
+            scheduler.attachFiber(fiber, body);
+        }).
+        join({
+            iterationCount: 0,
+            childFiberDidJoin(child, scheduler) {
+                delegate?.childFiberDidJoin?.call(delegate, child, scheduler);
+                this.iterationCount += 1;
+                const fiber = child.parent;
+                if (delegate?.repeatShouldEnd?.call(delegate, this.iterationCount, fiber, scheduler)) {
+                    return;
+                }
+                scheduler.attachFiber(fiber, body);
+            }
+        });
+    if (typeof f === "function") {
+        f(body);
+        return this;
+    }
+    return body;
+};
+
 // Set a value in its original scope, or in the fiber’s own scope if it does
 // not appear in any scope.
 ScheduledFiber.prototype.setOriginalValue = function(name, value) {
