+++ lib/util.js
@@ -10,7 +10,7 @@ export const K = x => () => x;
 export const clamp = (x, min, max) => Math.min(Math.max(x, min), max);
 
 // Extend an object by creating a new one with additional properties.
-export const extend = (x, ...props) => Object.assign(Object.create(x), ...props);
+export const extend = (x, ...props) => Object.assign(x ? Object.create(x) : {}, ...props);
 
 // Test whether a function is declared as async.
 export const isAsync = f => Object.getPrototypeOf(f) === Object.getPrototypeOf(async function() {});
